fix(install.sh): Clone repo with autocrlf=false settings
On some (windows) systems, if no ~/.gitconfig file, autocrlf settings were set to true, causing initial checkout to report every text file with modified line endings. Added 'autocrlf=false' to the clone command. It is already in the .gitconfig that gets cloned so should not be a problem after the initial clone.